Added an optimization step after invoking stoc
To be fast, the stochastic search does not work very hard at
optimization. It just finds a basic block that works, and then does a
quick dead code elimination on each iteration of the search to get rid
of around 85% of the crud.

So that means it can find less-than-ideal code sequences. For example if
we do the following:

    strop --arch mos6502 --function add15 --search stoc --in x --out y

to find a function that calculated X + 4 and leaves the result in Y, it
might generate this:

	sec
	txa
	adc #12
	tay
	iny
	iny

pretty stupid, moving X to the accumulator, adding 13 there (12 plus
carry), only to move it again to Y and then increment twice. So what
happens now, is we do a stochastic search towards more optimal code. And
that may produce the following:

	sec
	txa
	adc #14
	tay

Still not idiomatic, but it's just as good.