do not try and manipulate basic block past the end of the vector
This has come about because prog.len() now computes the length of the
program in bytes, rather than the length of the program in number of
instructions. So some of the mutations that can happen (insert
instruction at offset x, etc) now has had the chance of having x greater
than the possible maximum. This was always a software error, but it has
only manifested in a bug since we've started accounting for
variable-length instruction sets.